+++ src/main/java/org/sopt/confeti/domain/ticketvendor/application/dto/response/TicketVendorDtos.java
@@ -1,4 +1,4 @@
-package org.sopt.confeti.domain.ticketvendor.application.dto;
+package org.sopt.confeti.domain.ticketvendor.application.dto.response;
 
 import java.util.List;
